backend/requirements.txt
flask>=3.0
gunicorn>=21.2
supabase>=2.0
python-dotenv>=1.0
flask-cors>=4.0
PyJWT>=2.8
cryptography>=41.0
requests>=2.31
langchain-groq>=0.2
langchain-core>=0.3
PyGithub>=2.1
